Extract UI stuff from the overmap class into a separate file
- don't expose some of the functions anymore;
- astyle the new files.
